Add repo_type support `MANUAL`
`MANUAL` is just for 'localhost-admins' with full access to server and does not need any repos. It could be used for empty project and automate actions with ansible modules. But it could be used for manual repos with out-storage for projects and manual tests for ansible playbooks and other, or fast rewrite files and roles.