Synced hooks.
Refactoring and fix some problems in hooks. First of all hooks were
changed to synced execution. It is nessesary for good integrations.
Second, hooks for executions moved to ansible execution class because
this place is better and there is place on real execution.
Third, add some logging for better debug problems.